third_party/WebKit/LayoutTests/screen_orientation/lockOrientation-bad-argument-expected.txt

Validates that screen.lockOrientation() returns false on bad input

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS screen.orientation.type is "portrait-primary"
PASS screen.lockOrientation('invalid-orientation') threw exception TypeError: Failed to execute 'lockOrientation' on 'Screen': parameter 1 ('invalid-orientation') is not a valid enum value..
PASS screen.orientation.type is "portrait-primary"
PASS screen.lockOrientation(null) threw exception TypeError: Failed to execute 'lockOrientation' on 'Screen': parameter 1 ('null') is not a valid enum value..
PASS screen.orientation.type is "portrait-primary"
PASS screen.lockOrientation(undefined) threw exception TypeError: Failed to execute 'lockOrientation' on 'Screen': parameter 1 ('undefined') is not a valid enum value..
PASS screen.orientation.type is "portrait-primary"
PASS screen.lockOrientation(123) threw exception TypeError: Failed to execute 'lockOrientation' on 'Screen': parameter 1 ('123') is not a valid enum value..
PASS screen.orientation.type is "portrait-primary"
PASS screen.lockOrientation(window) threw exception TypeError: Failed to execute 'lockOrientation' on 'Screen': parameter 1 ('[object Window]') is not a valid enum value..
PASS screen.orientation.type is "portrait-primary"
PASS screen.lockOrientation(['portrait-primary', 'landscape-primary']) threw exception TypeError: Failed to execute 'lockOrientation' on 'Screen': parameter 1 ('portrait-primary,landscape-primary') is not a valid enum value..
PASS screen.orientation.type is "portrait-primary"
PASS screen.lockOrientation() threw exception TypeError: Failed to execute 'lockOrientation' on 'Screen': 1 argument required, but only 0 present..
PASS screen.orientation.type is "portrait-primary"
PASS successfullyParsed is true

TEST COMPLETE
